Update ee-backup-script.sh
I learned about `for loops` and put it to use here to help solve my first idea. The script runs fine, and backs up the files, but when it goes to remove the .sql file from the original location that it dropped it (webroot), it also deletes it from the backup location. I don't yet know how to solve this.