Update ee-backup-script.sh
I got some help from an individual on a coding Discord server. We went through and reformatted the code and he explained how it was wrong, and why I really shouldn't clutter it up with colours. I was having a lot of trouble with it not grabbing the server names properly and easy engine would fail when trying to backup the db file. Decided instead to print each site itself and just have lines that can be uncommented to allow for more than one site backup on the same box. Still struggling with the script deleting the .sql file from BOTH the original location (which is what we want) and from the backup location (not what we want).